Publish photos

+++ src/helpers/telegram.ts
@@ -11,6 +11,8 @@ import {
   Navigation
 } from '../interfaces/app.js'
 import { UserGender, User, UserFull } from '../interfaces/user.js'
+import { Topic } from '../interfaces/topic.js'
+import { Photo } from '../interfaces/photo.js'
 import {
   USER_NICK_REGEXP,
   USER_GENDERS,
@@ -35,11 +37,36 @@ export const wizardNextStep = (
 export const blankNavigation = (): Navigation => {
   const navigation: Navigation = {
     messageId: null,
+    updatable: false,
     currentPage: 0,
     totalPages: 0
   }
 
   return navigation
+}
+
+export const resetNavigation = (navigation: Navigation) => {
+  navigation.updatable = false
+  navigation.currentPage = 0
+  navigation.totalPages = 0
+}
+
+export const navigationNextPage = (navigation: Navigation) => {
+  if (
+    navigation.currentPage >= 1 &&
+    navigation.currentPage < navigation.totalPages
+  ) {
+    navigation.currentPage = navigation.currentPage + 1
+  }
+}
+
+export const navigationPrevPage = (navigation: Navigation) => {
+  if (
+    navigation.currentPage > 1 &&
+    navigation.currentPage <= navigation.totalPages
+  ) {
+    navigation.currentPage = navigation.currentPage - 1
+  }
 }
 
 export const blankMembership = (): Membership => {
@@ -103,6 +130,10 @@ export const sureSceneSessionRegister = (ctx: AppContext): Partial<Register> => 
   return register
 }
 
+export const dropSceneSessionRegister = (ctx: AppContext) => {
+  delete ctx.scene.session.register
+}
+
 export const initSceneSessionChangeAvatar = (ctx: AppContext) => {
   ctx.scene.session.changeAvatar = {} as Partial<ChangeAvatar>
 }
@@ -116,6 +147,10 @@ export const sureSceneSessionChangeAvatar = (ctx: AppContext): Partial<ChangeAva
   return changeAvatar
 }
 
+export const dropSceneSessionChangeAvatar = (ctx: AppContext) => {
+  delete ctx.scene.session.changeAvatar
+}
+
 export const initSceneSessionChangeAbout = (ctx: AppContext) => {
   ctx.scene.session.changeAbout = {} as Partial<ChangeAbout>
 }
@@ -127,6 +162,44 @@ export const sureSceneSessionChangeAbout = (ctx: AppContext): Partial<ChangeAbou
   }
 
   return changeAbout
+}
+
+export const dropSceneSessionChangeAbout = (ctx: AppContext) => {
+  delete ctx.scene.session.changeAbout
+}
+
+export const initSceneSessionNewPhoto = (ctx: AppContext) => {
+  ctx.scene.session.newPhoto = {} as Partial<NewPhoto>
+}
+
+export const sureSceneSessionNewPhoto = (ctx: AppContext): Partial<NewPhoto> => {
+  const newPhoto = ctx.scene.session.newPhoto
+  if (newPhoto === undefined) {
+    throw new Error(`context scene session newPhoto lost`)
+  }
+
+  return newPhoto
+}
+
+export const dropSceneSessionNewPhoto = (ctx: AppContext) => {
+  delete ctx.scene.session.newPhoto
+}
+
+export const initSceneSessionDeletePhoto = (ctx: AppContext, photoId: number) => {
+  ctx.scene.session.deletePhoto = photoId
+}
+
+export const sureSceneSessionDeletePhoto = (ctx: AppContext): number => {
+  const deletePhoto = ctx.scene.session.deletePhoto
+  if (deletePhoto === undefined) {
+    throw new Error(`context scene session deletePhoto lost`)
+  }
+
+  return deletePhoto
+}
+
+export const dropSceneSessionDeletePhoto = (ctx: AppContext) => {
+  delete ctx.scene.session.deletePhoto
 }
 
 export const isUserGender = (userGender: unknown): userGender is UserGender => {
@@ -233,37 +306,71 @@ export const keyboardProfileMenu = () => {
   return Markup.inlineKeyboard([
     [Markup.button.callback('Редактировать аватар', 'profile-change-avatar')],
     [Markup.button.callback('Редактировать о себе', 'profile-change-about')],
-    [Markup.button.callback('Вернуться в меню', 'profile-return-main')]
+    [Markup.button.callback('Вернуться в главное меню', 'profile-back')]
   ])
 }
 
-export const keyboardChangeAvatarCancel = () => {
+export const keyboardChangeAvatarMenu = () => {
   return Markup.inlineKeyboard([
     Markup.button.callback('Отмена', 'change-avatar-back')
   ])
 }
 
-export const keyboardChangeAboutCancel = () => {
+export const keyboardChangeAboutMenu = () => {
   return Markup.inlineKeyboard([
     Markup.button.callback('Отмена', 'change-about-back')
   ])
 }
 
-export const keyboardPhotoMenu = () => {
+export const keyboardPhotoMenu = (photo: Photo, navigation: Navigation) => {
+  const navButtons = []
+
+  if (navigation.currentPage !== 1) {
+    navButtons.push(Markup.button.callback('< Предыдущее', 'photo-prev'))
+  }
+
+  navButtons.push(Markup.button.callback('Перейти', 'photo-goto'))
+
+  if (navigation.currentPage !== navigation.totalPages) {
+    navButtons.push(Markup.button.callback('Следующее >', 'photo-next'))
+  }
+
   return Markup.inlineKeyboard([
-    [
-      Markup.button.callback('Предыдущее', 'photo-prev'),
-      Markup.button.callback('Перейти', 'photo-goto'),
-      Markup.button.callback('Следующее', 'photo-next')
-    ],
-    [Markup.button.callback('Удалить', 'photo-delete')],
-    [Markup.button.callback('В главное меню', 'photo-return-main')]
+    navButtons,
+    [Markup.button.callback('Удалить', `photo-delete-${photo.id}`)],
+    [Markup.button.callback('Добавить', 'photo-new')],
+    [Markup.button.callback('Вернуться в главное меню', 'photo-back')]
   ])
 }
 
-export const keyboardNewPhotoCancel = () => {
+export const keyboardPhotoBlank = () => {
   return Markup.inlineKeyboard([
-    Markup.button.callback('Отмена', 'new-photo-cancel')
+    [Markup.button.callback('Добавить', 'photo-new')],
+    [Markup.button.callback('Вернуться в главное меню', 'photo-back')]
+  ])
+}
+
+export const keyboardNewPhotoPhoto = () => {
+  return Markup.inlineKeyboard([
+    Markup.button.callback('Отмена', 'new-photo-photo-back')
+  ])
+}
+
+export const keyboardNewPhotoTopics = (topics: Topic[]) => {
+  const topicsButtons = topics.map((topic) => {
+    return [Markup.button.callback(topic.name, `new-photo-topic-${topic.id}`)]
+  })
+
+  return Markup.inlineKeyboard([
+    ...topicsButtons,
+    [Markup.button.callback('Отмена', 'new-photo-topic-back')]
+  ])
+}
+
+export const keyboardNewPhotoConfirm = () => {
+  return Markup.inlineKeyboard([
+    Markup.button.callback('Отмена', 'new-photo-confirm-back'),
+    Markup.button.callback('Продолжить', 'new-photo-confirm-next')
   ])
 }
 
@@ -271,7 +378,7 @@ export const removeLastMessage = async (
   ctx: AppContext,
   navigation: Navigation
 ): Promise<void> => {
-  if (navigation.messageId !== null) {
+  if (navigation.messageId !== null && !navigation.updatable) {
     await ctx.deleteMessage(navigation.messageId)
 
     navigation.messageId = null
@@ -450,7 +557,7 @@ export const replyProfileMenu = async (
   navigation.messageId = message.message_id
 }
 
-export const replyChangeAvatar = async (
+export const replyChangeAvatarMenu = async (
   ctx: AppContext,
   authorize: User,
   navigation: Navigation
@@ -459,13 +566,13 @@ export const replyChangeAvatar = async (
 
   const message = await ctx.replyWithMarkdownV2(
     `Загрузи новый аватар`,
-    keyboardChangeAvatarCancel()
+    keyboardChangeAvatarMenu()
   )
 
   navigation.messageId = message.message_id
 }
 
-export const replyChangeAbout = async (
+export const replyChangeAboutMenu = async (
   ctx: AppContext,
   authorize: User,
   navigation: Navigation
@@ -474,7 +581,7 @@ export const replyChangeAbout = async (
 
   const message = await ctx.replyWithMarkdownV2(
     `Расскажи о себе`,
-    keyboardChangeAboutCancel()
+    keyboardChangeAboutMenu()
   )
 
   navigation.messageId = message.message_id
@@ -489,7 +596,7 @@ export const replyChangeAboutWrong = async (
 
   const message = await ctx.replyWithMarkdownV2(
     `Некорректный ввод, попробуй еще раз`,
-    keyboardChangeAboutCancel()
+    keyboardChangeAboutMenu()
   )
 
   navigation.messageId = message.message_id
@@ -498,15 +605,104 @@ export const replyChangeAboutWrong = async (
 export const replyPhotoMenu = async (
   ctx: AppContext,
   authorize: User,
+  navigation: Navigation,
+  photos: Photo[]
+): Promise<void> => {
+  removeLastMessage(ctx, navigation)
+
+  if (photos.length > 0) {
+    navigation.totalPages = photos.length
+
+    if (navigation.currentPage < 1) {
+      navigation.currentPage = 1
+    } else if (navigation.currentPage > navigation.totalPages) {
+      navigation.currentPage = navigation.totalPages
+    }
+
+    const navigationIndex = navigation.currentPage - 1
+
+    const photo = photos[navigation.currentPage - 1]
+    if (photo === undefined) {
+      throw new Error(`can't get user photo by index`)
+    }
+
+    const extra = keyboardPhotoMenu(photo, navigation)
+
+    extra.caption = photo.description
+
+    if (navigation.updatable) {
+      await ctx.editMessageMedia(
+        {
+          type: 'photo',
+          media: photo.tgFileId,
+        },
+        extra
+      )
+    } else {
+      navigation.updatable = true
+
+      const message = await ctx.sendPhoto(photo.tgFileId, extra)
+
+      navigation.messageId = message.message_id
+    }
+  } else {
+    const message = await ctx.replyWithMarkdownV2(
+      `У вас нет опубликованных фото\n` +
+      `Отправьте мне фото для публикации\n` +
+      `Вы можете загрузить 3 фото в течении 24 часов`,
+      keyboardPhotoBlank()
+    )
+
+    console.dir(keyboardPhotoBlank())
+
+    navigation.messageId = message.message_id
+
+    resetNavigation(navigation)
+  }
+}
+
+export const replyNewPhotoPhoto = async (
+  ctx: AppContext,
+  authorize: User,
   navigation: Navigation
 ): Promise<void> => {
   removeLastMessage(ctx, navigation)
 
-  const extra = keyboardPhotoMenu()
+  const message = await ctx.replyWithMarkdownV2(
+    `Загрузи фотограйию`,
+    keyboardNewPhotoPhoto()
+  )
 
-  extra.caption = `Тест тест тест`
+  navigation.messageId = message.message_id
+}
 
-  const message = await ctx.sendPhoto(USER_AVATAR_UNKNOWN, extra)
+export const replyNewPhotoTopics = async (
+  ctx: AppContext,
+  authorize: User,
+  navigation: Navigation,
+  topics: Topic[]
+): Promise<void> => {
+  removeLastMessage(ctx, navigation)
+
+  const message = await ctx.replyWithMarkdownV2(
+    `Выбери раздел`,
+    keyboardNewPhotoTopics(topics)
+  )
+
+  navigation.messageId = message.message_id
+}
+
+export const replyNewPhotoConfirm = async (
+  ctx: AppContext,
+  authorize: User,
+  navigation: Navigation,
+): Promise<void> => {
+  removeLastMessage(ctx, navigation)
+
+  const message = await ctx.replyWithMarkdownV2(
+    `Опубликовать фото?`,
+    keyboardNewPhotoConfirm()
+  )
 
   navigation.messageId = message.message_id
 }
